+++ app/Http/Controllers/FactureController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\enum\PermissionsEnum;
 use App\Mail\InvoiceMail;
 use App\Models\Commande;
 use App\Models\Facture;
 use App\Models\Payement;
+use App\Models\User;
 use Generator;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -13,43 +15,74 @@ use Illuminate\Support\Facades\Mail;
 
 class FactureController extends Controller
 {
+
     /**
-     * Display a listing of the resource.
+     * Affiche la liste de toutes les factures
+     *
+     * @return \Illuminate\Contracts\View\View
      */
     public function index()
     {
+        $user = User::find(Auth::user()->id) ;
+        if (!$user->can(PermissionsEnum::GERER_FACTURES->value)) {
+
+           abort(403, 'Vous n\'avez pas la permission de gérer les factures.');
+       }
         $factures = Facture::all();
         return view('facture.index', compact('factures'));
     }
 
+    /**
+     * Affiche la liste des payements pour  une facture
+     *
+     * @param Facture $facture
+     * @return \Illuminate\Contracts\View\View
+     */
     public function listePayement(Facture $facture){
+        $user = User::find(Auth::user()->id) ;
+        if (!$user->can(PermissionsEnum::GERER_FACTURES->value)) {
+
+           abort(403, 'Vous n\'avez pas la permission de gérer les factures.');
+       }
         $payements = $facture->payemants;
 
         return view('facture.listePayment' , compact('payements'  , 'facture'));
     }
 
+
     /**
-     * Show the form for creating a new resource.
+     * Affiche le formulaire de création d'un payement
+     *
+     * @param Facture $facture
+     * @return \Illuminate\Contracts\View\View
      */
     public function createPayment(Facture $facture)
     {
+        $user = User::find(Auth::user()->id) ;
+        if (!$user->can(PermissionsEnum::GERER_FACTURES->value)) {
+
+           abort(403, 'Vous n\'avez pas la permission de gérer les payements.');
+       }
         return view('facture.create', compact('facture'));
     }
 
-    //Enregistrer un payement à la commande
 
 
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Facture  $facture
-     * @return \Illuminate\Http\Response
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Models\Facture $facture
+     * @return \Illuminate\Http\RedirectResponse
      */
-
     public function storePayement(Request $request , Facture $facture){
         //dd( $facture);
+        $user = User::find(Auth::user()->id) ;
+        if (!$user->can(PermissionsEnum::GERER_FACTURES->value)) {
+
+           abort(403, 'Vous n\'avez pas la permission de gérer les payements.');
+       }
         if ($facture->montantRestant>0) {
 
             if ($request->montantPaye > $facture->montantRestant) {
@@ -81,11 +114,22 @@ class FactureController extends Controller
 
     }
 
+
     /**
      * Store a newly created resource in storage.
+     *
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Models\Commande $commande
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request , Commande $commande)
     {
+
+        $user = User::find(Auth::user()->id) ;
+        if (!$user->can(PermissionsEnum::GERER_FACTURES->value)) {
+
+           abort(403, 'Vous n\'avez pas la permission de gérer les factures.');
+       }
         //dd($request->all() , $commande);
 
         $factue = Facture::create([
@@ -102,7 +146,18 @@ class FactureController extends Controller
         return redirect()->route('facture.index');
     }
 
+    /**
+     * Envoie une facture par email au client lié à cette facture
+     *
+     * @param \App\Models\Facture $facture
+     * @return \Illuminate\Http\RedirectResponse
+     */
     public function envoyer(Facture $facture){
+        $user = User::find(Auth::user()->id) ;
+        if (!$user->can(PermissionsEnum::GERER_FACTURES->value)) {
+
+           abort(403, 'Vous n\'avez pas la permission de gérer les factures.');
+       }
 
         Mail::to($facture->commande->devis->client->email)->send(new InvoiceMail($facture));
 
@@ -111,42 +166,45 @@ class FactureController extends Controller
     }
 
 
+    /**
+     * Génère la facture liée à la commande $commande
+     *
+     * @param \App\Models\Facture $facture
+     *
+     * @return \Illuminate\Contracts\View\View
+     */
     public function generer(Facture $facture){
+        $user = User::find(Auth::user()->id) ;
+        if (!$user->can(PermissionsEnum::GERER_FACTURES->value)) {
+
+           abort(403, 'Vous n\'avez pas la permission de gérer les factures.');
+       }
         $items = $facture->commande->commandeItems;
         $commande = $facture->commande;
-        return view('facture.generer', compact('facture' , 'commande' , 'items'));
+        return view('facture.genererFacture', compact('facture' , 'commande' , 'items'));
 
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
-    }
+
+
+
 
     /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
+     * Supprime la facture $facture
+     *
+     * @param \App\Models\Facture $facture
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(Facture $facture)
     {
+
+        $user = User::find(Auth::user()->id) ;
+        if (!$user->can(PermissionsEnum::SUPPRIMER_FACTURE->value)) {
+
+           abort(403, 'Vous n\'avez pas la permission de supprimer les factures.');
+       }
+
+        $facture->payemants()->delete();
         $facture->delete();
         return redirect()->route('facture.index')->with('success' , 'Facture supprimer avec succès');
     }
